Updated Mesh2D.cs
This is the updated version of Mesh2D.cs with the method for Draw implemented. Test game has been updtaed as well to test the draw method. Currently throghing an error during build that is saying that it is out of the scope of the array. Points to line 38 in Vertex.cs